+++ packages/sdk/src/evm/core/classes/index.ts
@@ -15,12 +15,7 @@ export * from "./erc-721-with-quantity-signature-mintable";
 export * from "./erc-721-burnable";
 export * from "./erc-721-standard";
 export * from "./erc-1155";
-export * from "./erc-1155-enumerable";
-export * from "./erc-1155-mintable";
-export * from "./erc-1155-batch-mintable";
 export * from "./erc-1155-signature-mintable";
-export * from "./erc-1155-burnable";
-export * from "./erc-1155-lazymintable";
 export * from "./erc-1155-standard";
 export * from "./marketplace-direct";
 export * from "./marketplace-auction";
